+++ microservices/user_service/repository/user_repository.go
@@ -6,10 +6,10 @@ import (
 
 // UserRepository defines the interface for user data operations
 type UserRepository interface {
-	FindByID(id int64) (*User, error)
-	ExistsByID(id int64) (bool, error)
+	FindByID(id uint) (*User, error)
+	ExistsByID(id uint) (bool, error)
 	ExistsByEmail(email string) (bool, error)
 	Create(user *User) (*User, error)
 	Update(user *User) (*User, error)
-	Delete(id int64) error
+	Delete(id uint) error
 }
